[LiNK]HiFi/Stereo Review - Stereo Review
Stereo Review was an American magazine first published in 1958 by Ziff-Davis. It was one of a small group of magazines then available for the individual interested in high fidelity (the other major ones being Audio and High Fidelity) It featured a blend of record and equipment reviews, articles on music and musicians, and articles on technical issues and advice. Beginning in 1958 as HiFi & Music Review, the name changed to HiFi Review in 1959. Then it became HiFi/Stereo Review in 1961 to reflect the growing use of stereophonic technology in recordings and broadcasts. In 1968 it became Stereo Review, reflecting the broad shift to stereophonic reproduction and simplifying the title. In the late 1980s, the magazine was acquired by CBS Magazines  and in 1989 it absorbed High Fidelity magazine.